Support mercurial

+++ fish/functions/grm.fish
@@ -1,1 +1,5 @@
-git rm $argv
+if _is_hg
+    hg remove $argv
+else
+    git rm $argv
+end
